requirements.txt
line-bot-sdk==2.3.0
requests==2.28.1
python-dotenv==0.21.0
flask==2.2.2
Werkzeug==2.2.2
